requirements.txt
aiogram==2.25.1
Django==4.1.9
djongo==1.3.6
fastapi==0.97.0
loguru==0.7.0
mplfinance==0.12.9b7
numpy==1.24.3
pandas==2.0.2
pymongo==3.12.1
python-decouple==3.8
requests==2.31.0
scipy==1.10.1
TA-Lib==0.4.26
tinkoff-investments==0.2.0b50
uvicorn==0.22.0